remove native build tag and simplify platform separation
