+++ cmd/finch/main_native.go
@@ -1,38 +1,85 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build linux || native
+//go:build linux
 
 package main
 
 import (
+	"fmt"
+	"io"
+
 	"github.com/spf13/afero"
+	"github.com/spf13/cobra"
 
 	"github.com/runfinch/finch/pkg/command"
 	"github.com/runfinch/finch/pkg/config"
 	"github.com/runfinch/finch/pkg/flog"
+	"github.com/runfinch/finch/pkg/lima/wrapper"
 	"github.com/runfinch/finch/pkg/path"
+	"github.com/runfinch/finch/pkg/support"
 	"github.com/runfinch/finch/pkg/system"
+	"github.com/runfinch/finch/pkg/version"
 )
 
-func nerdctlCmdCreator(ecc command.Creator, logger flog.Logger, fp path.Finch, finchRootPath string) command.NerdctlCmdCreator {
-	return command.NewNerdctlCmdCreator(ecc,
+var newApp = func(
+	logger flog.Logger,
+	fp path.Finch,
+	fs afero.Fs,
+	fc *config.Finch,
+	stdOut io.Writer,
+	_,
+	finchRootPath string,
+	ecc command.Creator,
+) *cobra.Command {
+	usage := fmt.Sprintf("%v <command>", finchRootCmd)
+	rootCmd := &cobra.Command{
+		Use:           usage,
+		Short:         "Finch: open-source container development tool",
+		SilenceUsage:  true,
+		SilenceErrors: true,
+		Version:       version.Version,
+	}
+	// TODO: Decide when to forward --debug to the dependencies
+	// (e.g. nerdctl for container commands and limactl for VM commands).
+	rootCmd.PersistentFlags().Bool("debug", false, "running under debug mode")
+	rootCmd.PersistentPreRunE = func(cmd *cobra.Command, _ []string) error {
+		// running commands under debug mode will print out debug logs
+		debugMode, _ := cmd.Flags().GetBool("debug")
+		if debugMode {
+			logger.SetLevel(flog.Debug)
+		}
+		return nil
+	}
+
+	ncc := command.NewNerdctlCmdCreator(ecc,
 		logger,
 		fp.NerdctlConfigFilePath(finchRootPath),
 		fp.BuildkitSocketPath(finchRootPath),
 		fp.FinchDependencyBinDir(),
 		system.NewStdLib(),
 	)
-}
-
-func configureNerdctl(fs afero.Fs) error {
-	return config.NewNerdctlApplier(
-		nil,
+	lima := wrapper.NewLimaWrapper()
+	supportBundleBuilder := support.NewBundleBuilder(
+		logger,
 		fs,
-		"",
-		"",
-		"",
-		"",
-		nil,
-	).Apply("")
+		support.NewBundleConfig(fp, finchRootPath),
+		fp,
+		ecc,
+		ncc,
+		lima,
+	)
+
+	// append nerdctl commands
+	allCommands := initializeNerdctlCommands(ncc, ecc, logger, fs, fc)
+	// append finch specific commands
+	allCommands = append(allCommands,
+		newVersionCommand(ncc, logger, stdOut),
+		newSupportBundleCommand(logger, supportBundleBuilder, ncc),
+		newGenDocsCommand(rootCmd, logger, fs, system.NewStdLib()),
+	)
+
+	rootCmd.AddCommand(allCommands...)
+
+	return rootCmd
 }
