+++ cmd/finch/support_bundle_native.go
@@ -1,10 +1,10 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build linux || native
+//go:build linux
 
 package main
 
-func (gsa *generateSupportBundleAction) assertVMExists() error {
+func (gsa *generateSupportBundleAction) canCreateBundle() error {
 	return nil
 }
