+++ e2e/vm/lifecycle_remote_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build (darwin || windows) && !native
+//go:build darwin || windows
 
 package vm
