+++ pkg/command/nerdctl_native.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build linux || native
+//go:build linux
 
 package command
