+++ pkg/command/nerdctl_remote_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build (darwin || windows) && !native
+//go:build darwin || windows
 
 // Package command_test is not named as command to avoid circular dependency (command <-> mocks).
 package command_test
