+++ pkg/config/nerdctl_config_applier.go
@@ -1,21 +1,28 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build darwin || windows
+
 package config
 
 import (
 	"errors"
 	"fmt"
 	"path"
+	"path/filepath"
+	"strings"
 
 	toml "github.com/pelletier/go-toml"
+	"github.com/pkg/sftp"
 	"github.com/spf13/afero"
+	"github.com/spf13/afero/sftpfs"
 
 	"github.com/runfinch/finch/pkg/fssh"
 )
 
 const (
-	nerdctlNamespace = "finch"
+	nerdctlNamespace      = "finch"
+	nerdctlRootfulCfgPath = "/etc/nerdctl/nerdctl.toml"
 )
 
 type nerdctlConfigApplier struct {
@@ -52,6 +59,70 @@ func NewNerdctlApplier(
 	}
 }
 
+func addLineToBashrc(fs afero.Fs, profileFilePath string, profStr string, cmd string) (string, error) {
+	if !strings.Contains(profStr, cmd) {
+		profBufWithCmd := fmt.Sprintf("%s\n%s", profStr, cmd)
+		if err := afero.WriteFile(fs, profileFilePath, []byte(profBufWithCmd), 0o600); err != nil {
+			return "", fmt.Errorf("failed to write to profile file: %w", err)
+		}
+		return profBufWithCmd, nil
+	}
+	return profStr, nil
+}
+
+// updateEnvironment adds variables to the user's shell's environment. Currently it uses ~/.bashrc because
+// Bash is the default shell and Bash will not load ~/.profile if ~/.bash_profile exists (which it does).
+// ~/.bash_profile sources ~/.bashrc, so ~/.bashrc is currently the best place to define additional variables.
+// The [GNU docs for Bash] explain how these files work together in more details.
+// The default location of DOCKER_CONFIG is ~/.docker/config.json. This config change sets the location to
+// ~/.finch/config.json, but from the perspective of macOS (/Users/<user>/.finch/config.json).
+// The reason that we don't set environment variables inside /root/.bashrc is that the vars inside it are
+// not able to be picked up even if we specify `sudo -E`. We have to switch to root user in order to access them, while
+// normally we would access the VM as the regular user.
+// For more information on the variable, see the registry nerdctl docs.
+//
+// [GNU docs for Bash]: https://www.gnu.org/software/bash/manual/html_node/Bash-Startup-Files.html
+//
+// [registry nerdctl docs]: https://github.com/containerd/nerdctl/blob/master/docs/registry.md
+
+func updateEnvironment(fs afero.Fs, fc *Finch, finchDir, homeDir, limaVMHomeDir string) error {
+	cmdArr := []string{
+		`export DOCKER_CONFIG="$FINCH_DIR"`,
+		"[ -L /usr/local/bin/docker-credential-ecr-login ] " +
+			`|| sudo ln -s "$FINCH_DIR"/cred-helpers/docker-credential-ecr-login /usr/local/bin/`,
+		`[ -L /root/.aws ] || sudo ln -fs "$AWS_DIR" /root/.aws`,
+	}
+
+	awsDir := fmt.Sprintf("%s/.aws", homeDir)
+
+	if *fc.VMType == "wsl2" {
+		cmdArr = append([]string{
+			fmt.Sprintf(`FINCH_DIR="$(/usr/bin/wslpath '%s')"`, finchDir),
+			fmt.Sprintf(`AWS_DIR="$(/usr/bin/wslpath '%s')"`, awsDir),
+		}, cmdArr...)
+	} else {
+		cmdArr = append([]string{
+			fmt.Sprintf(`FINCH_DIR=%s`, finchDir),
+			fmt.Sprintf(`AWS_DIR=%s`, awsDir),
+		}, cmdArr...)
+	}
+
+	profileFilePath := fmt.Sprintf("%s/.bashrc", limaVMHomeDir)
+	profBuf, err := afero.ReadFile(fs, profileFilePath)
+	if err != nil {
+		return fmt.Errorf("failed to read config file %q: %w", profileFilePath, err)
+	}
+	profStr := string(profBuf)
+	for _, element := range cmdArr {
+		profStr, err = addLineToBashrc(fs, profileFilePath, profStr, element)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
 // updateNerdctlConfig reads from the nerdctl config and updates values.
 func updateNerdctlConfig(fs afero.Fs, homeDir string, rootless bool) error {
 	nerdctlRootlessCfgPath := fmt.Sprintf("%s/.config/nerdctl/nerdctl.toml", homeDir)
@@ -76,7 +147,7 @@ func updateNerdctlConfig(fs afero.Fs, homeDir string, rootless bool) error {
 	var cfg Nerdctl
 	cfgBuf, err := afero.ReadFile(fs, cfgPath)
 	if err != nil {
-		return fmt.Errorf("updateEnvironment %q: %w", cfgPath, err)
+		return fmt.Errorf("failed to read config file %q: %w", cfgPath, err)
 	}
 
 	if err := toml.Unmarshal(cfgBuf, &cfg); err != nil {
@@ -97,24 +168,68 @@ func updateNerdctlConfig(fs afero.Fs, homeDir string, rootless bool) error {
 	return nil
 }
 
-// Apply updates the nerdctl config.
-func (nca *nerdctlConfigApplier) Apply(remoteAddr string) error {
-	nFs, err := nca.nerdctlFs(nca.fs, nca.privateKeyPath, remoteAddr)
-	if err != nil {
-		return fmt.Errorf("failed to get nerdctlFs: %w", err)
+func getLimaHomeDir(localFs afero.Fs, remoteFs afero.Fs, limaInstanceDir string, cfg *Finch) (string, error) {
+	var envB []byte
+	var envErr error
+
+	if *cfg.VMType == "wsl2" {
+		envB, envErr = afero.ReadFile(localFs, filepath.Join(limaInstanceDir, "cidata", "lima.env"))
+	} else {
+		envB, envErr = afero.ReadFile(remoteFs, "/mnt/lima-cidata/lima.env")
 	}
 
-	homeDir, err := getHomeDir(nca.fs, nFs, nca.limaInstancePath, nca.fc, nca.homeDir)
+	if envErr != nil {
+		return "", envErr
+	}
+
+	var user, home string
+	lines := strings.Split(string(envB), "\n")
+	for _, l := range lines {
+		if strings.Contains(l, "LIMA_CIDATA_USER") {
+			user = strings.Split(l, "=")[1]
+		} else if strings.Contains(l, "LIMA_CIDATA_HOME") {
+			home = strings.Split(l, "=")[1]
+		}
+	}
+
+	if home != "" {
+		return home, nil
+	}
+
+	return "/home/" + user + ".linux", nil
+}
+
+// Apply gets SSH and SFTP clients and uses them to update the nerdctl config.
+func (nca *nerdctlConfigApplier) Apply(remoteAddr string) error {
+	user := "root"
+	sshCfg, err := fssh.NewClientConfig(nca.fs, user, nca.privateKeyPath)
+	if err != nil {
+		return fmt.Errorf("failed to create ssh client config: %w", err)
+	}
+
+	sshClient, err := nca.dialer.Dial("tcp", remoteAddr, sshCfg)
+	if err != nil {
+		return fmt.Errorf("failed to setup ssh client: %w", err)
+	}
+
+	sftpClient, err := sftp.NewClient(sshClient)
+	if err != nil {
+		return fmt.Errorf("failed to setup sftp client: %w", err)
+	}
+
+	sftpFs := sftpfs.New(sftpClient)
+
+	limaHomeDir, err := getLimaHomeDir(nca.fs, sftpFs, nca.limaInstancePath, nca.fc)
 	if err != nil {
 		return fmt.Errorf("failed to get lima home dir: %w", err)
 	}
 
 	// rootless hardcoded to false for now to match our finch.yaml file
-	if err := updateNerdctlConfig(nFs, homeDir, false); err != nil {
+	if err := updateNerdctlConfig(sftpFs, limaHomeDir, false); err != nil {
 		return fmt.Errorf("failed to update the nerdctl config file: %w", err)
 	}
 
-	if err := updateEnvironment(nFs, nca.fc, nca.finchDir, homeDir, homeDir); err != nil {
+	if err := updateEnvironment(sftpFs, nca.fc, nca.finchDir, nca.homeDir, limaHomeDir); err != nil {
 		return fmt.Errorf("failed to update the user's .profile file: %w", err)
 	}
 	return nil
