+++ pkg/dependency/vmnet/binaries_unix_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build darwin && !native
+//go:build darwin
 
 // Ensures that the binaries required for networking are installed in a privileged location.
 // More information here: https://github.com/lima-vm/socket_vmnet
