+++ pkg/dependency/vmnet/update_network_lima_config_unix.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build darwin && !native
+//go:build darwin
 
 package vmnet
