+++ pkg/dependency/vmnet/vmnet_unix.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build darwin && !native
+//go:build darwin
 
 // Package vmnet handles installation and configuration of dependencies needed for Lima's managed networking
 // and port-forwarding to work, with minimal user interaction.
