+++ pkg/disk/disk_darwin.go
@@ -1,8 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build darwin && !native
-// +build darwin,!native
+//go:build darwin
 
 package disk
